+++ db/statement.go
@@ -25,7 +25,7 @@ type stmtWithErrWrapping struct {
 func (s stmtWithErrWrapping) Exec(ctx context.Context, args ...any) error {
 	err := s.Stmt.Exec(ctx, args...)
 	if err != nil {
-		return sqldb.WrapErrorWithQuery(err, s.PreparedQuery(), args, s.fmt)
+		return sqldb.WrapErrorWithQueryIfConfigured(err, s.PreparedQuery(), args, s.fmt)
 	}
 	return nil
 }
@@ -33,7 +33,7 @@ func (s stmtWithErrWrapping) Exec(ctx context.Context, args ...any) error {
 func (s stmtWithErrWrapping) Query(ctx context.Context, args ...any) sqldb.Rows {
 	rows := s.Stmt.Query(ctx, args...)
 	if rows.Err() != nil {
-		return sqldb.NewErrRows(sqldb.WrapErrorWithQuery(rows.Err(), s.PreparedQuery(), args, s.fmt))
+		return sqldb.NewErrRows(sqldb.WrapErrorWithQueryIfConfigured(rows.Err(), s.PreparedQuery(), args, s.fmt))
 	}
 	return rows
 }